+++ src/main/java/helper/StringHelper.java
@@ -1,4 +1,4 @@
-package calculator;
+package helper;
 
 import java.util.Random;
 
@@ -13,6 +13,12 @@ public class StringHelper {
 
     public static void throwExceptionIfNullOrEmpty(String input) {
         if (isNullOrEmpty(input)) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+    public static void throwExceptionNullOrLengthGreaterThan(String input, int length) {
+        if (input == null || input.length() > length) {
             throw new IllegalArgumentException();
         }
     }
