+++ src/test/java/racing/RacingTest.java
@@ -1,5 +1,6 @@
 package racing;
 
+import helper.NumberHelper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
